requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
mysqlclient==2.2.6
mysql-connector-python==9.1.0
Flask-Migrate==4.0.7
Flask-Script==2.0.6
google-generativeai==0.8.3
